Remove duplication of message delivery logic
Before this commit, we had several places where host.rs was checking
time vs envelope instructions to determine if a udp, syn or stream
message was visible. Now, this logic is encapsulated in `Link`. Or,
messages still on the network are now.. still on the network.

While this commit fixes the duplication, streams are completely broken
as they need to be reimplemented on top of the new structure. The udp
based tests pass.